Move agent specific otel components to agent/otel package
